+++ src/test/java/com/github/blutorange/maven/plugin/closurecompiler/test/MinifyMojoTest.java
@@ -290,6 +290,11 @@ public class MinifyMojoTest {
   }
 
   @Test
+  public void testIsolatePolyfills() throws Exception {
+    runMinifyAndAssertDirContent("isolatepolyfills");
+  }
+
+  @Test
   public void testSkip() throws Exception {
     runMinifyAndAssertDirContent("skip");
   }